sidebar using plain js. tangible assets reworked

+++ hmfinplan/database/migrations/2021_10_15_080544_create_real_estates_table.php
@@ -20,9 +20,9 @@ class CreateRealEstatesTable extends Migration
             $table->string('type', 25)->nullable();
             $table->string('desc', 50)->nullable();
             $table->date('purchase_yr')->nullable();
-            $table->integer('purchase_cost')->nullable();
+            $table->decimal('purchase_cost', 18, 2)->nullable();
             $table->decimal('expct_growth_rt', 5, 2)->nullable();
-            $table->decimal('current_val', 18, 4)->nullable();
+            $table->decimal('current_val', 18, 2)->nullable();
 		    $table->string('area')->nullable();
             $table->tinyInteger('status')->nullable();
             $table->timestamps();
